🚛 rename command from pruefi to ahb
