+++ src/kohlrahbi/changehistory/command.py
@@ -7,8 +7,8 @@ from pathlib import Path
 import click
 from maus.edifact import EdifactFormatVersion
 
+from kohlrahbi.ahb.command import check_python_version, validate_path
 from kohlrahbi.changehistory import scrape_change_histories
-from kohlrahbi.pruefi.command import check_python_version, validate_path
 
 
 @click.command()
