+++ src/kohlrahbi/read_functions.py
@@ -12,7 +12,6 @@ from docx.text.paragraph import Paragraph  # type:ignore[import]
 
 from kohlrahbi.ahbtable.ahbsubtable import AhbSubTable
 from kohlrahbi.ahbtable.ahbtable import AhbTable
-from kohlrahbi.changehistory.changehistorytable import ChangeHistoryTable
 from kohlrahbi.logger import logger
 from kohlrahbi.seed import Seed
 
@@ -208,31 +207,3 @@ def log_pruefi_not_found(pruefi):
     Logs that the Prüfi was not found in the provided document.
     """
     logger.warning("Prüfi '%s' was not found in the provided document.", pruefi)
-
-
-def is_change_history_table(table: Table) -> bool:
-    """
-    Checks if the given table is change history table.
-    """
-    # in the document 'Entscheidungsbaum-DiagrammeundCodelisten-informatorischeLesefassung3.5_99991231_20240401.docx'
-    # I got the error "IndexError: list index out of range", I am not sure which table caused the error
-    try:
-        return table.cell(row_idx=0, col_idx=0).text.strip() == "Änd-ID"
-    except IndexError:
-        return False
-
-
-def get_change_history_table(document: Document) -> Optional[ChangeHistoryTable]:
-    """
-    Reads a docx file and extracts the change history.
-    Returns None if no such table was found.
-    """
-
-    # Iterate through the whole word document
-    logger.info("🔁 Start iterating through paragraphs and tables")
-    for item in get_all_paragraphs_and_tables(parent=document):
-        if isinstance(item, Table) and is_change_history_table(table=item):
-            change_history_table = ChangeHistoryTable.from_docx_change_history_table(docx_table=item)
-            return change_history_table
-
-    return None
