+++ unittests/test_input_checks.py
@@ -1,6 +1,6 @@
 import pytest  # type:ignore[import]
 
-from kohlrahbi.pruefi import get_valid_pruefis
+from kohlrahbi.ahb import get_valid_pruefis
 
 
 @pytest.mark.parametrize(
